test(session): move migration regression to module suite
